fix(search): match clickable link targets
Some PDFs expose correct URL or destination targets while their visible link text is misencoded or split by glyph mappings. Searching link targets gives agents a reliable way to find those references and returns the clickable bbox without forcing links output into the public result.
