+++ skills/pdfvision/SKILL.md
@@ -5,7 +5,7 @@ description: "Extract text, metadata, per-page density signals, structural layou
 
 # pdfvision
 
-[pdfvision](https://github.com/yamadashy/pdfvision) extracts text + metadata + per-page density signals from any PDF, with opt-in OCR (`--ocr`, including OCR word boxes when tesseract returns layout), layout reconstruction (`--layout`), page anomaly warnings (`pages[].warnings[]` from layout geometry, glyph-noise, numeric-table, scan/OCR-layer, and image-box signals), raster image bounding boxes (`--image-boxes`, including image-bearing pattern fills), vector drawing counts (`vectorCount`), per-text-item geometry (`--geometry`), PNG rendering (`--render`, optionally sized with `--render-scale` and cropped with `--render-region`), and per-page text/form-field/annotation/OCR search with bbox (`--search`, hits ride into `--render-region` for one-pipeline find-then-zoom). Cached by content hash, so the second read of the same PDF returns in ~30 ms.
+[pdfvision](https://github.com/yamadashy/pdfvision) extracts text + metadata + per-page density signals from any PDF, with opt-in OCR (`--ocr`, including OCR word boxes when tesseract returns layout), layout reconstruction (`--layout`), page anomaly warnings (`pages[].warnings[]` from layout geometry, glyph-noise, numeric-table, scan/OCR-layer, and image-box signals), raster image bounding boxes (`--image-boxes`, including image-bearing pattern fills), vector drawing counts (`vectorCount`), per-text-item geometry (`--geometry`), PNG rendering (`--render`, optionally sized with `--render-scale` and cropped with `--render-region`), and per-page text/form-field/link/annotation/OCR search with bbox (`--search`, hits ride into `--render-region` for one-pipeline find-then-zoom). Cached by content hash, so the second read of the same PDF returns in ~30 ms.
 
 ## Prerequisite
 
@@ -87,7 +87,7 @@ The default extraction is enough for most native-text PDFs (papers, exports from
 | Hand the page to a vision model | `--render` + `--render-output <dir>` | Multimodal flows. Density Overview already flagged the page as low-text |
 | Shrink / enlarge the rendered PNG | `--render-scale <n>` (default 2, bounds `(0, 4]`) | 1×: half-size payload, fine for most agentic-vision dispatch. 3×+: capture chart / fine-print detail |
 | Zoom into a sub-rectangle of one page | `--render-region <x,y,w,h>` | Agent already saw a suspect block via `--layout` / `warnings[]` and only wants the visual confirmation of that bbox, not the whole page. PDF points, top-left origin, single-page only (errors if `--pages` resolves to multiple). Composes with `--render-scale`; on rotated pages, `pages[].rotation` explains why the cropped PNG's visible width/height may be swapped even though the input bbox stays in page coordinates |
-| Find every occurrence of a string with bbox | `--search <query>` (repeatable; `--search-regex` / `--search-case-sensitive` modifiers) | The agent's "where does this term appear?" question. Returns `pages[N].matches[*]` with span/word/widget/annotation-level bbox so the bbox feeds straight into `--render-region` for a follow-up visual zoom — one-pipeline find-then-zoom, no second pass. Markdown also renders a per-page `Search matches` table when search ran; use JSON/XML/TOON when a downstream tool needs to consume coordinates directly. Literal substring by default, case-insensitive, NFKC-aware (so `"fi"` matches the U+FB01 ligature), and CJK-aware enough that `科学` can match display-spaced `科 学`. Compact table-header rows can match phrase queries across adjacent column labels while keeping broad prose columns separated. Also searches visible text/choice form field values (match carries `source: 'formField'`; comb text widgets narrow to matching cells when available), visible FreeText annotation contents (`source: 'annotation'`), and OCR text when `--ocr` is on (match carries `source: 'ocr'`, using OCR word boxes when present and supplementing from full `ocr.text` when word reconstruction misses one or more occurrences); duplicate OCR hits already covered by non-OCR matches are suppressed. |
+| Find every occurrence of a string with bbox | `--search <query>` (repeatable; `--search-regex` / `--search-case-sensitive` modifiers) | The agent's "where does this term appear?" question. Returns `pages[N].matches[*]` with span/word/widget/link/annotation-level bbox so the bbox feeds straight into `--render-region` for a follow-up visual zoom — one-pipeline find-then-zoom, no second pass. Markdown also renders a per-page `Search matches` table when search ran; use JSON/XML/TOON when a downstream tool needs to consume coordinates directly. Literal substring by default, case-insensitive, NFKC-aware (so `"fi"` matches the U+FB01 ligature), and CJK-aware enough that `科学` can match display-spaced `科 学`. Compact table-header rows can match phrase queries across adjacent column labels while keeping broad prose columns separated. Also searches visible text/choice form field values (match carries `source: 'formField'`; comb text widgets narrow to matching cells when available), clickable link targets (`source: 'link'`), visible FreeText annotation contents (`source: 'annotation'`), and OCR text when `--ocr` is on (match carries `source: 'ocr'`, using OCR word boxes when present and supplementing from full `ocr.text` when word reconstruction misses one or more occurrences); duplicate OCR hits already covered by non-OCR matches are suppressed. |
 | Skip the on-disk cache | `--no-cache` | Forced re-extraction. Default behaviour is cache-on |
 
 ## Detecting silent failures with the density Overview
