+++ skills/pdfvision/references/structured-output.md
@@ -591,10 +591,10 @@ interface SearchMatch {
   page: number;                // 1-based, mirrors PageResult.page
   query: string;               // verbatim source query
   queryIndex?: number;         // 0-based into the search array; omitted for single-query calls
-  bbox: { x, y, width, height }; // union bbox of contributing spans/words/widget/annotation
-  boxes: { x, y, width, height }[]; // per-span/word/widget/annotation bboxes
+  bbox: { x, y, width, height }; // union bbox of contributing spans/words/widgets/links/annotations
+  boxes: { x, y, width, height }[]; // per-span/word/widget/link/annotation bboxes
   text: string;                // matched substring in the same form as pages[].text (NFKC when normalize is on)
-  source: 'native' | 'formField' | 'annotation' | 'ocr'; // native/span, formField/widget, annotation, or OCR
+  source: 'native' | 'formField' | 'link' | 'annotation' | 'ocr'; // native/span, formField/widget, link target, annotation, or OCR
   context?: string;            // surrounding line text for human / LLM readability
 }
 ```
@@ -621,6 +621,7 @@ pdfvision doc.pdf -p <m.page> --render --render-region <m.bbox.x>,<m.bbox.y>,<m.
 - **Native text is searched at reconstructed visual-line level**. A query can cross pdf.js span / font-run boundaries on the same line (e.g. `"Hello World"` split into `Hello` + `World`) and returns a union `bbox` plus per-span `boxes[]`, but narrow column gutters are treated as line breaks so matches do not stitch the end of one column to the start of another. Multi-line phrase stitching is intentionally not modelled yet because the resulting region is usually too broad for visual zoom.
 - **Partial native-span matches are sliced inside the span bbox**: horizontal spans are sliced along x, while tall vertical/rotated spans are sliced along y so `--render-region` can zoom to the matched word instead of the whole line.
 - **Text/choice form field values are searched too**. Form-field matches come back with `source: 'formField'` and use the widget bbox, even when `--form-fields` was not requested for output; comb text widgets narrow matches to the matching cells when pdf.js exposes `maxLen`/comb appearance metadata. Choice fields search the selected option's visible display value when it differs from the exported value. Internal values that are not visible text, such as unchecked checkbox `Off` or hidden/noView widget values, are not searched.
+- **Link targets are searched too**. Link matches come back with `source: 'link'` and use the clickable link bbox, even when `--links` was not requested for output. This lets URL / destination / attachment-target searches succeed even when the visible link text is glyph-garbled.
 - **Visible FreeText annotation contents are searched too**. Annotation matches come back with `source: 'annotation'` and use the annotation bbox, even when `--annotations` was not requested for output. Sticky-note popup contents and other normally closed annotation comments are not searched.
 - **OCR text is searched too when `--ocr` is on**. OCR-derived matches come back with `source: 'ocr'`; when `ocr.words[]` is present, `bbox`/`boxes[]` use OCR word geometry in the same page-point coordinate system as native spans. If word-level reconstruction misses one or more occurrences, pdfvision supplements from full `ocr.text` with a page-level bbox so no-space scripts and OCR line-boundary differences still remain searchable. If native text, a form-field value, or visible annotation text already produced the same query/text hit on that page, the duplicate OCR hit is suppressed so the precise non-OCR bbox wins; OCR-only extra hits are still emitted.
